+++ src/rfdetr/export/main.py
@@ -22,7 +22,7 @@ from torchvision.transforms.v2 import Compose, Resize, ToDtype, ToImage
 
 from rfdetr.datasets.transforms import Normalize
 from rfdetr.export._onnx.exporter import export_onnx
-from rfdetr.export.tensorrt import trtexec
+from rfdetr.export._tensorrt import trtexec
 from rfdetr.models import build_model
 from rfdetr.utilities.distributed import get_rank
 from rfdetr.utilities.logger import get_logger
